+++ src/PDB/Sequences.jl
@@ -1,6 +1,19 @@
 # Functions to extract the sequences from PDB structures.
 # =======================================================
 
+
+"""
+    is_aminoacid(residue::PDBResidue)
+    is_aminoacid(residue_id::PDBResidueIdentifier)
+
+This function returns `true` if the PDB residue is an amino acid residue. It checks if the 
+residue's three-letter name exists in the `MIToS.Utils.THREE2ONE` dictionary, and 
+returns `false` otherwise.
+"""
+is_aminoacid(residue::PDBResidue) = is_aminoacid(residue.id)
+is_aminoacid(residue_id::PDBResidueIdentifier) = _is_aminoacid(residue_id.name)
+_is_aminoacid(residue_name::String) = haskey(THREE2ONE, residue_name)
+
 function _add_sequence!(buf, chains, key)
     seq = String(take!(buf))
     if haskey(chains, key)
@@ -15,11 +28,13 @@ end
         model::Union{String,Type{All}}=All, chain::Union{String,Type{All}}=All, 
         group::Union{String,Regex,Type{All}}=All) where N
 
-This function generates a dictionary where each key-value pair consists of a named tuple 
-(representing the `model` and `chain` identifiers) and a corresponding protein sequence. 
-All residues are included by default, but those that don't satisfy specified criteria 
-based on `model`, `chain`, and `group` keyword arguments are excluded. If a residue name 
-isn't located in the `MIToS.Utils.THREE2ONE` dictionary, the function triggers a warning.
+This function returns an `OrderedDict` where each key is a named tuple (containing the 
+model and chain identifiers), and each value is the protein sequence corresponding to 
+the modelled residues in those chains. Therefore, the obtained sequences do not contain 
+missing residues. All modelled residues are included by default, but those that don't 
+satisfy specified criteria based on the `model`, `chain`, or `group` keyword arguments 
+are excluded. One-letter residue names are obtained from the `MIToS.Utils.THREE2ONE` 
+dictionary for all residue names that return `true` for `is_aminoacid`.
 """
 function modelled_sequences(residue_list::AbstractArray{PDBResidue,N}; 
         model::Union{String,Type{All}}=All, chain::Union{String,Type{All}}=All, group::Union{String,Type{All}}=All) where N
@@ -27,9 +42,8 @@ function modelled_sequences(residue_list::AbstractArray{PDBResidue,N};
     buf = IOBuffer()
     first_residue = first(residue_list)
     key = (model=first_residue.id.model, chain=first_residue.id.chain)
-    warnings = Dict{String, Set{String}}()
     for res in residue_list
-        if !_is(res.id.model, model) || !_is(res.id.chain, chain) || !_is(res.id.group, group) || res.id.name == "HOH"
+        if !_is(res.id.model, model) || !_is(res.id.chain, chain) || !_is(res.id.group, group) || !is_aminoacid(res)
             continue
         end
         current_key = (model=res.id.model, chain=res.id.chain)
@@ -37,19 +51,8 @@ function modelled_sequences(residue_list::AbstractArray{PDBResidue,N};
             _add_sequence!(buf, chains, key)
             key = current_key
         end
-        if haskey(THREE2ONE, res.id.name)
-            write(buf, THREE2ONE[res.id.name])
-        else
-            if haskey(warnings, res.id.name)
-                push!(warnings[res.id.name], res.id.number)
-            else
-                warnings[res.id.name] = Set([res.id.number])
-            end
-        end
+        write(buf, THREE2ONE[res.id.name])
     end
     _add_sequence!(buf, chains, key)
-    for (name, numbers) in warnings
-        @warn("Residue name $(name) (residue numbers: $(join(numbers, ", "))) is not found in MIToS.Utils.THREE2ONE")
-    end
     chains
 end